Complete refactoring of the whole project
I rewrote more or less everything to make the code base cleaner, and to
make KeePassPHP much more flexible and easy to use as a library.
KeePassPHP now exposes two API : a high-level API to use KeePassPHP as a
.kdbx file manager (with key files, etc.); and a low-level API to use
kdbx files decryption/encryption and parsing utilities. For example, the
kdbx format used by KeePass is quite general can be used to encrypt and
embed any kind of data.