Use openssl if present (otherwise, fallback to mcrypt), and make key transformation as fast as possible
The OpenSSL PHP extension is faster and way better maintained than
mcrypt, so let's use it if possible. Also, a lot of time can be gained
by just moving the encryption rounds loop into the cipher class (to
avoid tens of thousands of calls to $cipher->encrypt), so I added
Cipher::encryptManyTimes for this case.